requirements.txt
# ============================================
# DEPENDÊNCIAS COMPLETAS - Python 3.12+
# Projeto: LSTM Stock Predictor
# ============================================

# ============================================
# Data Manipulation & Processing
# ============================================
numpy>=1.26.0,<2.0.0
pandas>=2.1.0,<3.0.0

# ============================================
# Data Collection
# ============================================
yfinance>=0.2.52
requests>=2.31.0
requests-cache>=1.2.0

# ============================================
# Machine Learning - Traditional
# ============================================
scikit-learn>=1.4.0
statsmodels>=0.14.1

# ============================================
# Machine Learning - Time Series
# ============================================
prophet>=1.1.5
pmdarima>=2.0.4

# ============================================
# Machine Learning - Deep Learning
# ============================================
tensorflow>=2.16.1
keras>=3.0.0

# ============================================
# API Framework
# ============================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-multipart>=0.0.9

# ============================================
# Dashboard & UI
# ============================================
streamlit>=1.31.0

# ============================================
# MLOps & Monitoring
# ============================================
mlflow>=2.10.0
dvc>=3.48.0
prometheus-client>=0.19.0
psutil>=5.9.8

# ============================================
# Visualization
# ============================================
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# ============================================
# Notebooks (Opcional para produção)
# ============================================
jupyter>=1.0.0
ipykernel>=6.29.0
ipywidgets>=8.1.0

# ============================================
# Utilities
# ============================================
python-dotenv>=1.0.0
joblib>=1.3.2
pyyaml>=6.0.1
loguru>=0.7.2

# ============================================
# Testing & Quality
# ============================================
pytest>=8.0.0
pytest-cov>=4.1.0

# ============================================
# NOTA: Para instalação minimal (apenas essencial):
# Descomente abaixo e use: pip install -r requirements-minimal.txt
# - numpy, pandas, yfinance, scikit-learn, fastapi, uvicorn
# ============================================

pytest-cov>=4.1.0
httpx>=0.26.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.29.0
ipywidgets>=8.1.0
